authentication and authorization

+++ frontend/src/components/WorkoutDetails.js
@@ -1,16 +1,34 @@
+//import { Link } from 'react-router-dom'
 
+//import formatDistanceToNow from 'date-fns/formatDistanceToNow'
+//import parseISO from 'date-fns/parseISO'
+import  { useWorkoutContext }  from "../hooks/useWorkoutContext"
+//import { useWorkoutContext } from '../hooks/useWorkoutContext'
+import useAuthContext from '../hooks/useAuthContext'
 const WorkoutDetails = ({workout})=>{
+    const { dispatch } = useWorkoutContext()
+    const { user } = useAuthContext()
 
     const handleDelete = async () => {
+        
+        if(!user){
+            return
+        }
+
         const response = await fetch('api/workouts/' + workout._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization' : `Bearer ${user.token}`
+            }
         })
-        const json = response.json()
+        const json = await response.json()
 
         if(response.ok){
-            console.log(json, 'deleted')
+            dispatch({type: 'DELETE_WORKOUT', payload: json})
         }
     }
+
+    //const createdAtDate = parseISO(workout.createdAt)
 
     return (
     <div className="workout-details">
@@ -18,9 +36,14 @@ const WorkoutDetails = ({workout})=>{
         <p><strong>Load (kg): </strong>{workout.load}</p>
         <p><strong>Reps: </strong>{workout.reps}</p>
         <p><strong>{workout.createdAt}</strong></p>
-        <span onClick={handleDelete}className='material-symbols-outlined'>delete</span>
+        <span onClick={handleDelete}className='material-symbols-outlined'>
+            delete</span>
+
+            {/* <span onClick={handleDelete}className='material-symbols-outlined'>
+            <Link to='/' reloadDocument>delete</Link></span> */}
     </div>
     )
 }
 
+//{formatDistanceToNow(new Date(workout.createdAt), {addSuffix: true})}
 export default WorkoutDetails
